ZSH: disable menucomplete option
If you've got a lots of files with the same prefix, e.g.

- my-spreadsheet-01.xlsx
- my-spreadsheet-02.xlsx
- my-spreadsheet-03.xlsx

I want to be able to delete them by typing e.g.

> rm my<TAB>*

But if you have this option turned on, the <TAB> will put the whole of
`my-spreadsheet-01.xlsx` on the line, which is means it doesn't work.